Extract ButtonsPresenter from the ArgoHelper
Extracting a class provides better interfaces and encapsulation which makes
testing and debugging easier.
