+++ app/views/catalog/_show_external_links.html.erb
@@ -13,5 +13,5 @@
     <li><small><%= render_index_info document %></small></li>
   </ul>
   <hr>
-  <%= render 'items/button_link_list', buttons: render_buttons(document, object) %>
+  <%= render 'items/button_link_list', buttons: buttons_presenter.buttons %>
 </div>
