+++ app/views/catalog/_show_sidebar.html.erb
@@ -1,2 +1,20 @@
-<%= render 'catalog/show_external_links', document: @document, object: @obj %>
+<% document = @document %>
+<div class="show_sidebar">
+  <h3 class="start-open">View in new window</h3>
+  <ul class='nav nav-stacked'>
+    <li><%= render_purl_link document %></li>
+    <% if document.catkey.present? %>
+    <li><%= render_searchworks_link document %></li>
+    <% end %>
+    <li><%= render_dor_link document %></li>
+    <li><%= link_to 'FoXML', File.join(Dor::Config.fedora.safeurl, "objects/#{document.id}/export?context=archive"), target: '_blank' %></li>
+    <li>
+      <%= link_to 'Solr document', solr_document_path(document, format: :json), target: '_blank' %>
+    </li>
+    <li><small><%= render_index_info document %></small></li>
+  </ul>
+  <hr>
+  <%= render 'items/button_link_list', buttons: @buttons_presenter.buttons %>
+</div>
+
 <%= render 'catalog/show_embargo_sidebar', document: @document %>
