+++ spec/views/catalog/_show_sidebar.html.erb_spec.rb
@@ -2,23 +2,24 @@
 
 require 'rails_helper'
 
-describe 'catalog/_show_external_links.html.erb' do
+describe 'catalog/_show_sidebar.html.erb' do
   let(:document) do
     SolrDocument.new(id: 'druid:abc123', catkey_id_ssim: 'catz')
   end
   let(:current_user) { mock_user }
-  let(:object) { instance_double(Dor::Item) }
+  let(:button_presenter) {  }
 
   before do
     config = Blacklight::Configuration.new
     allow(view).to receive(:blacklight_config).and_return(config)
     allow(view).to receive(:current_user).and_return(current_user)
     allow(controller).to receive(:current_user).and_return(current_user)
+    @document = document
+    @buttons_presenter = instance_double(ButtonsPresenter, buttons: [])
   end
 
-  it 'renders link list' do
-    expect(view).to receive(:render_buttons).with(document, object).and_return({})
-    render 'catalog/show_external_links', document: document, object: object
+  it 'renders a list of links' do
+    render
     expect(rendered).to have_css '.show_sidebar'
     expect(rendered).to have_css 'ul.nav.nav-stacked'
     expect(rendered).to have_css 'li', count: 6
